updated module name
I forked is to work-around the problem of needed to upgrade GRPC in our
storenet. The upstream fixed #3131 is need for CLOUD-8313 but requires GRPC
upgrade which is a bigger effort, CLOUD-7492.
This PR cherry-pick commit #3131 to our existing gocloud.dev hash.

Once, CLOUD-7492 completed, we should remove this fork to use upstream's repo.